refactor and switch to using refactored code
[ci skip-rust]

[ci skip-build-wheels]
